neo4j/requirements.txt
neo4j>=5.0.0
python-dotenv>=1.0.0
supabase>=2.0.0

